Update Hospital Scheduling System - Main build
(-)Removed first prototype input function as it will have to be modified to account for the linked lists and associated functions initialized in new code.
(+)Added the linked_list class, and initialized the schedules of three doctors as 'empty' linked lists i.e. linked lists with 24 elements to account for the 24 time slots, but without any  pre-existing appointment details.